requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
mysqlclient==2.1.0
sqlparse==0.4.2
djangorestframework==3.13.1
pytz==2022.1
django-session-timeout==0.1.0
six==1.16.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-timezone-field==6.0.1
psycopg2==2.9.6
python-crontab==3.0.0
python-dateutil==2.8.2
redis==5.0.1
